Reset everything to start all over again...
