+++ app/build.gradle.kts
@@ -6,17 +6,8 @@ plugins {
     // The Kotlin Plugin...
     alias(notation = libs.plugins.android.kotlin.plugin)
 
-    // The Security Plugin (AppSweep)...
-    alias(notation = libs.plugins.appsweep)
-
-    // The Compiler Plugin Developer (KSP)...
-    alias(notation = libs.plugins.ksp)
-
-    // The Google Plugins...
-    alias(notation = libs.plugins.gms.google.services)
-    alias(notation = libs.plugins.firebase.crashlytics)
-    alias(notation = libs.plugins.secrets.gradle.plugin)
-    alias(notation = libs.plugins.firebase.performance)
+    // The Compose Compiler Gradle Plugin...
+    alias(notation = libs.plugins.compose.compiler)
 
 }
 
@@ -58,9 +49,6 @@ android {
         compose = true
         buildConfig = true
     }
-    composeOptions {
-        kotlinCompilerExtensionVersion = libs.versions.composeVersion.get()
-    }
     packaging {
         resources {
             excludes += "/META-INF/{AL2.0,LGPL2.1}"
@@ -84,52 +72,8 @@ dependencies {
     implementation(dependencyNotation = libs.androidx.compose.material3)
     implementation(dependencyNotation = libs.lifecycle.runtime.compose)
 
-    // Firebase...
-    implementation(dependencyNotation = platform(libs.firebase.bom))
-    implementation(dependencyNotation = libs.firebase.cloud.firestore)
-    implementation(dependencyNotation = libs.firebase.analytics)
-    implementation(dependencyNotation = libs.firebase.cloud.messaging)
-    implementation(dependencyNotation = libs.firebase.performance)
-    implementation(dependencyNotation = libs.firebase.appcheck.playintegrity)
-
-    // The Navigation Library (Jetpack Compose)...
-    implementation(dependencyNotation = libs.androidx.navigation.compose)
-
-    // The Dependency Injection Library (Koin)...
-    implementation(dependencyNotation = libs.koin.androidx.compose)
-
-    // The Logging Library (Timber)...
-    implementation(dependencyNotation = libs.timber)
-
-    // The Memory Leak Detection Library (LeakCanary)...
-    debugImplementation(dependencyNotation = libs.leak.canary)
-
-    // The Local Storage Library (Room)...
-    implementation(dependencyNotation = libs.androidx.room.runtime)
-    implementation(dependencyNotation = libs.androidx.room.ktx)
-    "ksp"(dependencyNotation = libs.androidx.room.compiler)
-
-    // The Image Loading Library (Coil)...
-    implementation(dependencyNotation = libs.coil)
-
-    // The Kotlin Coroutines Library...
-    implementation(dependencyNotation = libs.kotlinx.coroutines.core)
-    implementation(dependencyNotation = libs.kotlinx.coroutines.android)
-
-    // The Gemini API...
-    implementation(dependencyNotation = libs.gemini.api)
-
-    // The Exoplayer Library (Media3)...
-    implementation(dependencyNotation = libs.androidx.media3.exoplayer)
-
     // The Testing Libraries...
     testImplementation(dependencyNotation = libs.junit)
-    testImplementation(dependencyNotation = libs.robolectric)
-    testImplementation(dependencyNotation = libs.mockK)
-    testImplementation(dependencyNotation = libs.kotlinx.coroutines.test)
-    testImplementation(dependencyNotation = libs.koin.test)
-    testImplementation(dependencyNotation = libs.koin.test.junit)
-    testImplementation(dependencyNotation = libs.androidx.room.testing)
     androidTestImplementation(dependencyNotation = libs.androidx.test.ext.junit)
     androidTestImplementation(dependencyNotation = libs.androidx.test.espresso.core)
     androidTestImplementation(dependencyNotation = platform(libs.androidx.compose.bom))
