--- app/src/main/java/emmanuelmuturia/carizma/car/domainlayer/repository/CarRepository.kt
@@ -1,14 +1,0 @@
-package emmanuelmuturia.carizma.car.domainlayer.repository
-
-import com.google.ai.client.generativeai.type.GenerateContentResponse
-import emmanuelmuturia.carizma.car.domainlayer.model.Car
-
-interface CarRepository {
-
-    suspend fun getCarById(carId: Int): Car?
-
-    suspend fun getResponse(car: Car?): GenerateContentResponse
-
-    suspend fun getResult(car: Car?): String?
-
-}
