--- app/src/main/java/emmanuelmuturia/carizma/commons/dependencyinjection/FirebaseKoinModule.kt
@@ -1,14 +1,0 @@
-package emmanuelmuturia.carizma.commons.dependencyinjection
-
-import com.google.firebase.Firebase
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.firestore
-import org.koin.dsl.module
-
-val firebaseKoinModule = module {
-
-    single<FirebaseFirestore> {
-        Firebase.firestore
-    }
-
-}
