--- app/src/main/java/emmanuelmuturia/carizma/commons/domainlayer/CarizmaState.kt
@@ -1,11 +1,0 @@
-package emmanuelmuturia.carizma.commons.domainlayer
-
-sealed interface CarizmaState<out T> {
-
-    data class Success<T>(val data: T): CarizmaState<T>
-
-    data class Error(val message: Throwable): CarizmaState<Nothing>
-
-    data object Loading : CarizmaState<Nothing>
-
-}
