--- app/src/main/java/emmanuelmuturia/carizma/commons/uilayer/components/CarizmaDummyCars.kt
@@ -1,59 +1,0 @@
-package emmanuelmuturia.carizma.commons.uilayer.components
-
-import emmanuelmuturia.carizma.car.domainlayer.model.Car
-
-val koenigseggList = List(size = 7) {
-    Car(
-        carId = 0,
-        carName = "Koenigsegg Agera R",
-        carImage = "https://images.unsplash.com/photo-1596906673164-fffb05b8333d?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Mnx8a29lbmlnc2VnZ3xlbnwwfHwwfHx8MA%3D%3D"
-    )
-}
-
-val astonMartinList = List(size = 7) {
-    Car(
-        carId = 1,
-        carName = "Aston Martin DB5",
-        carImage = "https://images.unsplash.com/photo-1678379315146-ed23821c3e21?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTB8fGFzdG9uJTIwbWFydGluJTIwZGI1fGVufDB8fDB8fHww"
-    )
-}
-
-val porscheList = List(size = 7) {
-    Car(
-        carId = 2,
-        carName = "Porsche 911",
-        carImage = "https://images.unsplash.com/photo-1578937227978-050be101e3a9?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Nnx8cG9yY2hlJTIwOTExfGVufDB8fDB8fHww"
-    )
-}
-
-val mercedesBenzList = List(size = 7) {
-    Car(
-        carId = 3,
-        carName = "Mercedes Benz (AMG)",
-        carImage = "https://images.unsplash.com/photo-1618843479313-40f8afb4b4d8?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8NHx8bWVyY2VkZXMlMjBiZW56fGVufDB8fDB8fHww"
-    )
-}
-
-val ferrariList = List(size = 7) {
-    Car(
-        carId = 4,
-        carName = "Ferrari LaFerrari",
-        carImage = "https://images.unsplash.com/photo-1617060219602-8cbf8f1eff8d?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Mnx8ZmVycmFyaSUyMGxhJTIwZmVycmFyaXxlbnwwfHwwfHx8MA%3D%3D"
-    )
-}
-
-val mcLarenList = List(size = 7) {
-    Car(
-        carId = 5,
-        carName = "McLaren P1",
-        carImage = "https://images.unsplash.com/photo-1516298103702-43b22eb200e8?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTB8fG1jbGFyZW4lMjBwMXxlbnwwfHwwfHx8MA%3D%3D"
-    )
-}
-
-val audiList = List(size = 7) {
-    Car(
-        carId = 6,
-        carName = "Audi R8",
-        carImage = "https://images.unsplash.com/photo-1603584173870-7f23fdae1b7a?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8M3x8YXVkaSUyMHI4fGVufDB8fDB8fHww"
-    )
-}
