--- app/src/main/java/emmanuelmuturia/carizma/commons/uilayer/theme/Color.kt
@@ -1,17 +1,0 @@
-package emmanuelmuturia.carizma.commons.uilayer.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val CarizmaOrange = Color(color = 0xFFFC460D)
-val CarizmaLightOrange = Color(color = 0xFFE8670B)
-val CarizmaWhite = Color(color = 0xFFFFFFFF)
-val CarizmaGray = Color(color = 0xFF3F3B3B)
-val CarizmaTranslucent = Color(color = 0x99FFFFFF)
