--- app/src/main/java/emmanuelmuturia/carizma/home/domainlayer/repository/HomeRepository.kt
@@ -1,10 +1,0 @@
-package emmanuelmuturia.carizma.home.domainlayer.repository
-
-import emmanuelmuturia.carizma.car.domainlayer.model.Car
-import kotlinx.coroutines.flow.Flow
-
-fun interface HomeRepository {
-
-    fun getCars(): Flow<List<Car>>
-
-}
