+++ app/src/main/java/emmanuelmuturia/carizma/main/activity/MainActivity.kt
@@ -7,9 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import androidx.navigation.compose.rememberNavController
-import emmanuelmuturia.carizma.commons.uilayer.theme.CarizmaTheme
-import emmanuelmuturia.carizma.navigation.navgraph.NavGraph
+import emmanuelmuturia.carizma.main.theme.CarizmaTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -23,7 +21,7 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    NavGraph(navController = rememberNavController())
+
                 }
             }
         }
