+++ app/src/main/java/emmanuelmuturia/carizma/main/theme/Type.kt
@@ -1,0 +1,5 @@
+package emmanuelmuturia.carizma.main.theme
+
+import androidx.compose.material3.Typography
+
+val Typography = Typography()
