--- app/src/main/java/emmanuelmuturia/carizma/player/domainlayer/repository/PlayerRepository.kt
@@ -1,22 +1,0 @@
-package emmanuelmuturia.carizma.player.domainlayer.repository
-
-import androidx.media3.exoplayer.ExoPlayer
-import emmanuelmuturia.carizma.car.domainlayer.model.Car
-
-interface PlayerRepository {
-
-    val player: ExoPlayer
-
-    suspend fun getCarById(carId: Int): Car?
-
-    suspend fun playCarAudio(carAudio: String)
-
-    suspend fun resumeCarAudio()
-
-    suspend fun pauseCarAudio()
-
-    suspend fun rewindCarAudio()
-
-    suspend fun fastForwardCarAudio()
-
-}
