--- app/src/main/java/emmanuelmuturia/carizma/profile/uilayer/ProfileScreen.kt
@@ -1,10 +1,0 @@
-package emmanuelmuturia.carizma.profile.uilayer
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun ProfileScreen() {
-
-
-
-}
