--- app/src/main/java/emmanuelmuturia/carizma/settings/uilayer/SettingsScreen.kt
@@ -1,10 +1,0 @@
-package emmanuelmuturia.carizma.settings.uilayer
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun SettingsScreen() {
-
-
-
-}
