+++ gradle/libs.versions.toml
@@ -1,71 +1,25 @@
 [versions]
 
 # The Kotlin Plugin...
-kotlinVersion = "1.9.23"
+kotlinVersion = "2.0.0"
 
 # The Gradle Plugin...
-gradleVersion = "8.4.1"
-
-# The Compose Compiler...
-composeVersion = "1.5.13"
+gradleVersion = "8.5.2"
 
 # The Android Libraries (Kotlin)...
 androidxCoreKtxVersion = "1.13.1"
-androidxLifecycleRuntimeKtx = "2.7.0"
+androidxLifecycleRuntimeKtx = "2.8.4"
 androidxSplashScreenVersion = "1.0.1"
 
 # The Jetpack Compose Libraries...
-androidxActivityCompose = "1.9.0"
-androidxComposeBom = "2024.05.00"
-lifecycleRuntimeComposeVersion = "2.7.0"
-
-# The Navigation Library (Jetpack Compose)...
-androidxNavigationComposeVersion = "2.7.7"
-
-# The Dependency Injection Library (Koin)...
-koinVersion = "3.6.0-wasm-alpha2"
-
-# The Logging Library...
-timberVersion = "5.0.1"
-
-# The Security Plugin (AppSweep)...
-appSweepGradlePluginVersion = "latest.release"
-
-# The Memory Leak Detection Library (LeakCanary)...
-leakCanaryVersion = "2.13"
-
-# The Compiler Plugin Developer Plugin (KSP)...
-kspVersion = "1.9.23-1.0.20"
-
-# The Local Storage Library (Room)...
-roomVersion = "2.6.1"
-
-# The Image Loading Library (Coil)...
-coilVersion = "2.6.0"
-
-# The Kotlin Coroutines...
-kotlinCoroutinesVersion = "1.8.0"
-
-# ExoPlayer (Media3)...
-media3ExoPlayerVersion = "1.3.1"
-
-# The Google Libraries & Plugins...
-googleServicesVersion = "4.4.1"
-firebaseCrashlyticsVersion = "2.9.9"
-secretsGradlePlugin = "2.0.1"
-firebasePerformanceVersion = "1.4.2"
-firebaseBOMVersion = "33.0.0"
-
-# Gemini API...
-geminiApiVersion = "0.6.0"
+androidxActivityCompose = "1.9.1"
+androidxComposeBom = "2024.06.00"
+lifecycleRuntimeComposeVersion = "2.8.4"
 
 # The Testing Libraries...
 junitVersion = "4.13.2"
-androidxTestExtJunitVersion = "1.1.5"
-androidxTestEspressoCore = "3.5.1"
-robolectricVersion = "4.12.1"
-mockKVersion = "1.13.10"
-kotlinxCoroutinesTestVersion = "1.8.1-Beta"
+androidxTestExtJunitVersion = "1.2.1"
+androidxTestEspressoCore = "3.6.1"
 
 [libraries]
 
@@ -84,54 +38,10 @@ androidx-compose-ui-tooling = { group = "androidx.compose.ui", name = "ui-toolin
 androidx-compose-material3 = { group = "androidx.compose.material3", name = "material3" }
 lifecycle-runtime-compose = { group = "androidx.lifecycle", name = "lifecycle-runtime-compose", version.ref = "lifecycleRuntimeComposeVersion" }
 
-# The Navigation Library (Jetpack Compose)...
-androidx-navigation-compose = { group = "androidx.navigation", name = "navigation-compose", version.ref = "androidxNavigationComposeVersion" }
-
-# The Dependency Injection Library (Koin)...
-koin-androidx-compose = { module = "io.insert-koin:koin-androidx-compose", version.ref = "koinVersion" }
-
-# The Logging Library (Timber)...
-timber = { group = "com.jakewharton.timber", name = "timber", version.ref = "timberVersion" }
-
-# The Memory Leak Detection Library (LeakCanary)...
-leak-canary = { group = "com.squareup.leakcanary", name = "leakcanary-android", version.ref = "leakCanaryVersion" }
-
-# The Local Storage Library (Room)...
-androidx-room-runtime = { group = "androidx.room", name = "room-runtime", version.ref = "roomVersion" }
-androidx-room-compiler = { group = "androidx.room", name = "room-compiler", version.ref = "roomVersion" }
-androidx-room-testing = { group = "androidx.room", name = "room-testing", version.ref = "roomVersion" }
-androidx-room-ktx = { group = "androidx.room", name = "room-ktx", version.ref = "roomVersion" }
-
-# The Image Loading Library (Coil)...
-coil = { group = "io.coil-kt", name = "coil-compose", version.ref = "coilVersion" }
-
-# The Kotlin Coroutines Library...
-kotlinx-coroutines-core = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core", version.ref = "kotlinCoroutinesVersion" }
-kotlinx-coroutines-android = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-android", version.ref = "kotlinCoroutinesVersion" }
-
-# The Exoplayer Library (Media3)...
-androidx-media3-exoplayer = { module = "androidx.media3:media3-exoplayer", version.ref = "media3ExoPlayerVersion" }
-
-# Firebase...
-firebase-bom = { module = "com.google.firebase:firebase-bom", version.ref = "firebaseBOMVersion" }
-firebase-cloud-firestore = { group = "com.google.firebase", name = "firebase-firestore" }
-firebase-analytics = { group = "com.google.firebase", name = "firebase-analytics" }
-firebase-cloud-messaging = { group = "com.google.firebase", name = "firebase-messaging" }
-firebase-performance = { group = "com.google.firebase", name = "firebase-perf" }
-firebase-appcheck-playintegrity = { group = "com.google.firebase", name = "firebase-appcheck-playintegrity" }
-
-# Gemini API...
-gemini-api = { module = "com.google.ai.client.generativeai:generativeai", version.ref = "geminiApiVersion" }
-
 # The Testing Libraries...
 junit = { group = "junit", name = "junit", version.ref = "junitVersion" }
 androidx-test-ext-junit = { group = "androidx.test.ext", name = "junit", version.ref = "androidxTestExtJunitVersion" }
 androidx-test-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "androidxTestEspressoCore" }
-robolectric = { group = "org.robolectric", name = "robolectric", version.ref = "robolectricVersion" }
-mockK = { group = "io.mockk", name = "mockk", version.ref = "mockKVersion" }
-kotlinx-coroutines-test = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-test", version.ref = "kotlinxCoroutinesTestVersion" }
-koin-test = { module = "io.insert-koin:koin-test", version.ref = "koinVersion" }
-koin-test-junit = { module = "io.insert-koin:koin-test-junit4", version.ref = "koinVersion" }
 androidx-compose-ui-test-manifest = { group = "androidx.compose.ui", name = "ui-test-manifest" }
 androidx-compose-ui-test-junit4 = { group = "androidx.compose.ui", name = "ui-test-junit4" }
 
@@ -143,14 +53,5 @@ android-gradle-plugin = { id = "com.android.application", version.ref = "gradleV
 # The Kotlin Plugin...
 android-kotlin-plugin = { id = "org.jetbrains.kotlin.android", version.ref = "kotlinVersion" }
 
-# The Security Plugin (AppSweep)...
-appsweep = { id = "com.guardsquare.appsweep", version.ref = "appSweepGradlePluginVersion" }
-
-# The Google Plugins...
-gms-google-services = { id = "com.google.gms.google-services", version.ref = "googleServicesVersion" }
-firebase-crashlytics = { id = "com.google.firebase.crashlytics", version.ref = "firebaseCrashlyticsVersion" }
-secrets-gradle-plugin = { id = "com.google.android.libraries.mapsplatform.secrets-gradle-plugin", version.ref = "secretsGradlePlugin" }
-firebase-performance = { id = "com.google.firebase.firebase-perf", version.ref = "firebasePerformanceVersion" }
-
-# The Compiler Plugin Developer Plugin (KSP)...
-ksp = { id = "com.google.devtools.ksp", version.ref = "kspVersion" }
+# The Compose Compiler Gradle Plugin...
+compose-compiler = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlinVersion" }
