+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,6 @@
-#Wed May 22 08:02:13 EAT 2024
+#Mon Aug 19 21:10:46 EAT 2024
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.6-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.7-bin.zip
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
